libcamera2: [WORKAROUND]enable SW JPEG encoder for merr_vv
BZ: 81671

This is an workaround just for merr_vv pre-Alpha targets. It should be
reverted later when HW JPEG encoder driver is ready.

Without this patch, the system may hang when exiting camera application,
capturing an image or switching between two camera sensors as described
in BZ: 80309.

This patch includes:
1. Disable HW JPEG encoder with USE_INTEL_JPEG := false
2. Disable the Intel Metabuffer for compile reasons.

Change-Id: I5415852148b88faeaa9f14e9952532e9d6775a04
Signed-off-by: Bin Han <bin.b.han@intel.com>
Reviewed-on: http://android.intel.com:8080/87721
Reviewed-by: Yang, Liang L <liang.l.yang@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>