libcamera2: debug: print count of preview frames when stopping
BZ: 81750

To debug problems in starting up ISP, print the number of preview
frames processed when stopping preview. This information already
exists in HAL, so we just need to print it.

To also cover cases where we do preview restarts, move reseting
of preview count from message handled to startPreviewCore().

Change-Id: I6caedb484ea09614c7f7f0504a0ffc5985ea43a5
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/87828
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>