libcamera2: single capture: optimizing static parameter processing
BZ: 82662

in the processing of static parameter, we are calling a lot
of unecessary code, which should be used only in video mode
and not in single capture.
This leads to extra stop and start ISP streaming in single capture.

This patch checks if we are in video mode or single capture mode
before calling the video aspect ratio check functions.

Change-Id: I7c1bea8995b2a968e193a226fe702232bfcf8f1c
Signed-off-by: Arnaud Mandy <arnaud.mandy@intel.com>
Reviewed-on: http://android.intel.com:8080/88390
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Jakonen, JarnoX <jarnox.jakonen@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>