libcamera2: make sure picture thread is flushed when stopping preview
BZ: 80087

The recent "libcamera2: host preview state from PreviewThread" patch
created a problem where picture thread flush was sometimes skipped
altogether: it was done done in stopPreviewCore due to the shot2shot
optimization (not really stopping preview), and neither when whole
camera stack was closed (as assumption was that stopPreviewCore
would do the cleanup).

The easy fix would be to just flush the picture thread in stopPreviewCore(),
but this breaks the flash sequence when started in continuous
preview mode (a flush of picture thread in a middle of capture
sequence loses application request state).

This patch fixes the problem by doing the picture thread flush by
default in in stopPreviewCore() and adding special-case handling
for the flash fallback. In the latter case, capture is not
in progress (we have just run the preflash sequence), so flush of
picture thread is not needed.

Change-Id: Ieeea155bbb3762707229d73f3d524894f19bfa34
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/88680
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>