libcamera2: do not skip snapshot frames in continuous mode
BZ: 83225

When burst is done continuous ISP mode (only for Time Nudge/Shift in
current code), the burst capture count must be configured beforehand
to ISP. Due to this, it's not possible to arbitrarily skip frames
during the capture sequence.

This patch disables the HAL-based burst skip logic and moves
the related common code to a single function.

Change-Id: If623899eba15593f36d30bceb2acc588bc1b05b2
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/89095
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>