REVERT "libcamera2: set the overlay orientation for CTP to 0"
BZ: 83675

This reverts commit d6694eb65118e2bdadf315cb431fc4922cb05dfb.
Using back CPU rotation for overlay support in CTP phone.
The real rotation needed is 90 degrees. When we put it to 0
we effectively let the Gfx to do the rotation for the Overlay
plane. This was not known to be possible at the time.
Now we now it is possible but it is not cost effective from
power consumption point of view.

Change-Id: I6a02375e7b696377c61eb8841905702939a8eeae
Reviewed-on: http://android.intel.com:8080/89230
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Tested-by: Patale, Pavan <pavan.patale@intel.com>
Reviewed-by: Patale, Pavan <pavan.patale@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>