libcamera2: modify GRALLOC flags for preview to allow overlay
BZ: 83978

With the newly added support on the HWC for normal NV12 format
now our traditional Gfx buffers can be rendered via overlay.
Only need a few changes in the GRALLOC flags.

After this change the terminology in the PreviewThread becomes
a bit confusing. Because the buffers managed by the GfxrPreviewHandler
can actually end up in the overlay.

So the main diference now between this handler and the overlay is who
does the rotation.
In case of the OverlayPreviewHandler the rotation is done in CPU with the
implicit memcopy that the rotation routine introduces

In case of the GfxPreviewHandler the rotation is done in GFx and the data
flow is 0-copy

Change-Id: I38823dbac4301d4b4af665ea0cc6ba885517d553
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/89769
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Tested-by: Patale, Pavan <pavan.patale@intel.com>
Reviewed-by: Patale, Pavan <pavan.patale@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>