libcamera2: downgrade log traces to level 2
BZ: 84944

with the new preview flow we have more threads polling and
the traces of v4l2_poll becomes quite verbose.
Downgrading this to level 2 to keep the log level one clean

Also in case were preview is doing a memcopy there is a log
level 1 trace that appears for each preview frame.

Change-Id: I1e07ae7ee62cf06ebbb50ae390dc03aaffc1ceeb
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/90046
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Patale, Pavan <pavan.patale@intel.com>
Tested-by: Patale, Pavan <pavan.patale@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>