libcamera2: apply sensor flip both for MAIN and PREVIEW device node
BZ: 85163

This patch addes the applySensorFlip() for
V4L2_PREVIEW_DEVICE(/dev/video2), the reason is that: when configuring
camera device to start preview, only the V4L2_PREVIEW_DEVICE will be
configured rathen than V4L2_MAIN_DEVICE(/dev/video0). So, libcamera2
needs to applySensorFlip() for preview case.

Without this patch, the issue found on Saltbay is that the flip setting
can not be applied for preview until when I pressed the capture key.

Change-Id: I92c426b0d865b818e8a57059f4452d0aa3320dbb
Signed-off-by: Bin Han <bin.b.han@intel.com>
Reviewed-on: http://android.intel.com:8080/90277
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Zheng, Hongjiang A <hongjiang.a.zheng@intel.com>
Reviewed-by: Zhang, Jie J <jie.j.zhang@intel.com>
Tested-by: Patale, Pavan <pavan.patale@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>