libcamera2: Introducing post capture processing thread
BZ: 83203

The new post capture processing thread is a generic thread
to off-load time consuming operations

It is meant for alogrithms like ULL, HDR, etc...

The control thread will off-load this processing
after the capture is taken.
Once the processing is completed the control thread will
be notified so that it can complete the jpeg encoding.

Change-Id: Iba9c6590fc25a1951843c5bb48e7c7d9887f15c4
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/91866
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>