libcamera2: fix recovery from preview starving
BZ: 86211

Previously added preview parallelization broke the recovery handling
of ISP timeouts because new ObserverThread commonly pauses in case of
error.

This change adds a retry logic that checks whether the case is that
driver is starving from preview buffers. When it is, we sleep for 200ms
and try again.

In other cases than starving, we more likely need a complete reset and
should handle the case through ControlThread.

Change-Id: Ief1895782315d24707c6addbb1609f3d6191046d
Signed-off-by: Teemu Tuominen <teemux.tuominen@intel.com>
Reviewed-on: http://android.intel.com:8080/91259
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Patale, Pavan <pavan.patale@intel.com>
Reviewed-by: Mandy, Arnaud <arnaud.mandy@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Tested-by: Patale, Pavan <pavan.patale@intel.com>
Reviewed-by: Gozalvez Herrero, Juan AntonioX <juan.antoniox.gozalvez.herrero@intel.com>
Tested-by: Gozalvez Herrero, Juan AntonioX <juan.antoniox.gozalvez.herrero@intel.com>