libcamera2: get camera configuration from xml file by according to board.
BZ: 84917

Issue:
Our customer wants to configure the camera settings without modifying any code.
So we need to move the camera configuration to file.
The file is different between boards.

Solution:
The xml format is the industry standard to store data.
So we store the camera configuration to the xml file.
They are different by boards.
The user could modify it to change the camera settings.

Depends on:
95511 95512

Change-Id: Ife4553d6a5309a1e028e5a0474e89f69329a288a
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/95220
Reviewed-by: Zheng, Hongjiang A <hongjiang.a.zheng@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: Kaikumaa, Timo <timo.kaikumaa@intel.com>
Reviewed-by: Seibel, Eric <eric.seibel@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>