libcamera2: hdr: release hdr when cancel capture
BZ: 87650

HDR enable flag is a parameter.
if during HDR capture a message is sent from application
without HDR enabled, The release of buffers will not be
done correctly, since the flag will be reset.
For example, during cancelling a capture.

This patch introduces a new flag, which is internal to
libcamera2 and is only true when hdr capture is in progress

Change-Id: Ia65f05ec0b8e8ed5275fe70de2b83701907102fb
Signed-off-by: Arnaud Mandy <arnaud.mandy@intel.com>
Reviewed-on: http://android.intel.com:8080/94599
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>