libcamera2: Unoptimized fallback NV12 SW rotation.
BZ: 91888

For overlay case we are doing the rotation in CPU

Current behaviour in case of a non optimized resolution
is requested is to avoid the rotation, this produces
a solid green frame.

We introduce now a fallback 90-degrees rotation routine
for NV12 that it is executed instead.

This will make that in case of non supported resoluton
we rotate with an unoptimal and produce a warning for every frame
ather than show a solid green frame

Change-Id: I3ae21bcd0416276be3690cf3d8f8417e2471d96d
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>