libcamera2: re-unify PreviewThread message handlers [1/2]
BZ: 91542

After the addition for NV21 support in the HWC and the support
for non-0-copy patch in the Gfx preview handler there is no
actual reason to have this level of abstraction in the preview thread

This patch set is split in 2 patches.
The first one removes the PrevewiMessageHandler abstraction and leaves only
the support for the GFx buffers.

The second one will add the overlay support again.

Change-Id: Id5bfc7432f7e5fbfedb3d66e4de8c6d960ac65ab
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>