libcamera2: re-unify PreviewTrhead message handlers [2/2]
BZ: 91542

In this second part of the patch set we re-introduce the support for
the overlay rendering.

We implement the logic to swap the dimensions of the preview window
and add the CPU rotation routines.

We also change the stride required for the rotation since we are using
now the plain NV12 format and not the NV12_VED.
NV12 requires only 64 alignment whereas
NV12_VED required 1024

Change-Id: I5563cc46c8b8282eaf2210539cbc734a1f1d22c1
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>