libcamera2: allocate snapshot buffers later to shorten L2P
BZ: 92632

This patch makes snapshot buffer allocation happen after preview
has started in order to shorten the launch to preview latency.

For continuous capture this have a more complex implication
that is:
The snapshot buffers v4l pool is no longer initialized before
start preview but after. This means that before each
startOfflineCapture we need to pass the new snapshot buffers
allocated by the PictureThread again to the ISP and also call
AtomISP::allocate() to update the v4l2 pool, like we do in old
online mode

We remove also the unnecessary preparation of the main and
postview devices during startContinuousPreview.  This is what
will be done before the startOfflineCapture from now on.

This provides flexibility of configuring initially a ring buffer
of certain size for capture, and later capture less or same
if we want to (this will be useful in ULL)

Change-Id: Ie74b74b101009131acde2b358124e935896a37a1
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/96551
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>