[PORT FROM MAIN] libcamera2: cached sensor data to shorten L2P
BZ: 98139

This patch stores the sensor data into a global variable so that it
doesn't need to be fetched from the sensor every time the preview is
started. This reduces the L2P time. NOTE: the memory that is
allocated during the first sensor data fetch into that global static
variable is intentionally not freed.

Change-Id: Iafe1846e49f10cd41b1ba951c879ad641bb59d41
Orig-Change-Id: Ida43776c6aee43e4f5a511e2f3f6bb197edfd4f4
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Reviewed-on: http://android.intel.com:8080/101138
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>