libcamera2: set V4L2 buffer cache flags
BZ: 93617

The V4L2 driver needs to know on a per-buffer basis whether the capture
buffers will be used by the host CPU, so the driver can perform
appropriate invalidation/flushing if needed.

The default behaviour of V4L2 is that driver should always invalidate.
This patch marks cases where HAL explicitly allocates non-cached memory
itself, or ISP output is routed directly to other hardware blocks
(graphics, jpeg encoder). This helps with performance as unnecessary
cache flushes are not done.

Specifically DMA operations of the Atom CSS are not cache coherent
(CSS2300/2400), so explicit invalidation is needed when the data is used
by host CPU.

Change-Id: I253830e5845cd5b7d928412436be4e3f7c6be716
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/97366
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>