libcamera2: wrong frame rate used to calculate shutter alignment
BZ: 94613

The AtomISP::getFrameRate() may return still mode FPS even when
AtomISP is configured for preview mode, and this caused the shutter
alignment calculation to produce different values in some use-case
sequences, which lead to preview start failures.

Fix the problem by using a hardcoded preview rate in the calculations.
We should ask this from kernel, but currently HAL must start preview
before kernel can provide the accurate frame information, and for
shutter lag alignment, we need the value before starting preview, so the
current interfaces are not suitable.

Change-Id: I090971cc8ce873b5aa055cd1a07fa2c021293e9c
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/97735
Reviewed-by: Mandy, Arnaud <arnaud.mandy@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Vtminen, LassiX <lassix.vtminen@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>