libcamera2: cleanup: remove dead code
BZ: 95019

Remove obsolete code not used by anybody
setSnapshotNumber used to be called to se the number
of snapshot numbers that where allocated by the AtomISP
class.

Nowadays the snapshot buffers are  pre-allocated by
the PictureThread so they can be attached to the JPEG
HW encoder context.
The buffers are passed to AtomISP via the call:
setSnapshotBuffers()

Change-Id: Ife4422e39a8c61f08ee2b4156ca5b8a70a2dedfa
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/97858
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>