libcamera2: check before re-allocating postview buffers
BZ: 92632

Problem:
After the previous optimization for Laucnh to Preview (L2P)
we now allocate the snapshot buffers before each capture.

The actual allocation of snapshot buffers is already done
we are only configuring the v4l2 buffers pool. Well we are actually
 allocating the shared buffers (4 bytes only)

But for postviews we were really freeing and allocating several frames
as big as preview.
This produced a regresion on the Shot to Shot (S2S) metric

Solution

We split the allocation of snaphots and postviews so that we
can check whether we really need new postviews or not.

Change-Id: Ib3259c30d7b10510068c0b799f3f1a5ea82c6021
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/98397
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>