[PORT FROM MAIN] libcamera2: cleanup preview state- and capture sequence-semantics
BZ: 73975

ControlThread's internal state has no longer reflected the real capture
state towards the public HAL API. This is due that ControlThread does not
transit its state from preview to capture when running in continuous-mode
and also due that the state transitions to capture always happen within
asynchronous message handler.

This change introduces ControlThread::mStillCaptureInProgress to simplify
the state needed for acting coherently with HAL API on sequences starting
from takePicture(). Semantic of previewEnabled() during
mStillCaptuerInProgress is true is separated from PreviewThread states.

Drops preview-keep-alive parameter extension and changes the naming
conventions of postview rendering to not to confuse with the removed
extension.

ControlThread is made responsible of hiding the preview and deciding
whether to pass the postview frames to display in different capture
modes. Postview frames get displayed regardless if preview state is
set to hidden.

PreviewThread::postview() now supports hide-request flag to be able
to hide the following preview stream after the postview is handled.
By default, hiding is requested with single capture done in continuous
mode. When postview of correct format is send to PreviewThread it gets
rendered also in hidden state.

Change-Id: I71c40d810c4bf4ec39f6891975c7a5fd41b817ed
Orig-Change-Id: Ie8ff7115b312ff1590e7bbb4a45c3598784f3e35
Signed-off-by: Teemu Tuominen <teemux.tuominen@intel.com>
Reviewed-on: http://android.intel.com:8080/100220
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/101751