libcamera2: fix potential memory leak of postview buffers
BZ: 96440

The recent optimization in change:
- Ib3259c30d7b10510068c0b799f3f1a5ea82c6021

did split the snaphot buffer de-allocation method
in two, one to free the snapshot buffers and
another one to free the postview.

That patch failed to ensure that the postview buffers were
free in the same situation where they used to.

So it is possible that in some occasions the postview buffers
 were left allocated.
This patch fixes that problem

Change-Id: Ia008aaec297fea5204e2434cb56f0b73e74bacad
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/98887
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Mandy, Arnaud <arnaud.mandy@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>