[PORT FROM MAIN] libcamera2: fix potential memory leak of postview buffers
BZ: 96440

The recent optimization in change:
- Ib3259c30d7b10510068c0b799f3f1a5ea82c6021

did split the snaphot buffer de-allocation method
in two, one to free the snapshot buffers and
another one to free the postview.

That patch failed to ensure that the postview buffers were
free in the same situation where they used to.

So it is possible that in some occasions the postview buffers
 were left allocated.
This patch fixes that problem

Change-Id: If250a65887c7ebfed3376d66e551f7c38e6a08c3
Orig-Change-Id: Ia008aaec297fea5204e2434cb56f0b73e74bacad
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>