libcamera2: return the correct ringbuffer size regardless of the ISP mode
BZ: 96446

Problem:
In cases where we fallback from continuous to normal capture
(like when taking images with flash) the AtomISP mode was left
in a state that made the query for ringbuffersize to return 1

This had the effect that we were allocating only 1 buffer when
we actually wanted more. Like when ULL is active.
and the the capture got stuck for lack of buffer.

Solution.
Remove the mode check in the query method

Change-Id: Ie91856e01de1a5b18c5bdb819cb9410a863f172e
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/99659
Reviewed-by: Niemimuukko, Ukri <ukri.niemimuukko@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>