[PORT FROM MAIN] libcamera2: return the correct ringbuffer size regardless of the ISP mode
BZ: 96446

Problem:
In cases wherewe fallback from continuous to normal capture
(like when taking images with flash) the AtomISP mode was left
 in a state that made the query for ringbuffersize to return 1

This had the effect that we were allocating only 1 buffer when
we actually wanted more. Like when ULL is active.
and the the capture got stuck for lack of buffer.

Solution.
Remove the mode check in the query method

Change-Id: I0fb8d448283c6fff5f6fef051b486c6b94dbf481
Orig-Change-Id: Ie91856e01de1a5b18c5bdb819cb9410a863f172e
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>