[PORT FROM MAIN] libcamera2: Add dark threshold to deactivate ULL in very dark scene
BZ: 96446

Current implementation of 3A logic to activate ULL had only
one threshold. This meant that once the scene was dark enough
ULL was activated and Flash never triggered

This patch introduces a second threshold to deactivate
ULL in case the scene is too dark. In those situations the flash
will be used.

Also the threshold logic is simplified. Until now we had different
threshold depending on whether we were going into the ULL range
or coming out from it. The intention was to avoid flickering of
the ULL activation decision.

This is removed by this patch and a simple range check is used.

Change-Id: Ie3b73bf11cdd1292090aeaaa42ca06f82d8b7c7f
Orig-Change-Id: I629e0878557f6b1a232d55c67d346d7bbe8787e2
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/99761
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>