[PORT FROM MAIN] libcamera2: drop flush and return buffers from preview backgrounding
BZ: 98938

When preview is left running in background, PreviewThread was flushed
and preview buffers in flow were returned to ISP. Flushing of buffers
was not done synchronously in PostProcThread leaving currently processed
buffer open to call IBufferOwner::returnBuffer() afterwards.

While other options exist, the buffers in flow do not need to be flushed at
all for the backgrounding. This change removes flushing and returning of buffers
from the use case and drops AtomISP::returnPreviewBuffers() originally added for
this purpose.

Change-Id: I37bcaadc64177f13abf6466d2eec7910609bb375
Orig-Change-Id: I5f4ba60c2863988a7aa5018a8f3b60e9bc06809d
Signed-off-by: Teemu Tuominen <teemux.tuominen@intel.com>
Reviewed-on: http://android.intel.com:8080/100683
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/101753