[PORT FROM R42]libcamera2: ULL: add method to retrieve input after usage
BZ: 97724

We add a new method to be able to retrieve the input buffers after
processing is done
This method is needed so that we can return the buffers to the pool
of buffers in ControlThread.

Since we are using the first input one as target destination for the
output of the algorithm we only return the others.

Change-Id: Ibdceca55f4d884079a5329669ff1e191eb9dc758
Orig-Change-Id: Ie08e32840c5083ad97997d622365270ace88ab7a
Reviewed-on: http://android.intel.com:8080/101487
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>