libcamera2: ULL: Use ApexSv values as activation threshold
BZ: 98129

For RHB we use AccuteLogic 3A that uses this value for exposure
This value is more reliable than manual iso.

Also disable ULL if Flash would be needed.
This removes the need for the other threshold we had before
Effectively it means that the flash activation will overrule the
3A activation of ULL.

Change-Id: I5b9102f4afb27c420a1c7f5c615e7a9a39fa44e9
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/101757
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>