libcamera2: initialize flipping variable
BZ: 99214

Problem: Moving to use XML instead of the old style platform
data left the flipping variable uninitialized.

This patch fixes the issue by initializing the variable properly.
Unfortunately the classes in PlatformData.h had to be moved around,
since the definition for the enum values was in the latter class.

Change-Id: I3d7cd9677ff965ce00b859da32102c96249f90c2
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Reviewed-on: http://android.intel.com:8080/100932
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Yang, Liang L <liang.l.yang@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>