libcamera2: fix for Teewinot 1024x768 recording/preview use case
BZ: 95781

This patch has the fix for Teewinot 1024x768 recording/preview use case.
The problem with this resolution is that the processing time requirement
for ISP (VFPP) is longer than sensor's vertical blanking time.

This patch implements video recording from preview buffers by swapping
the device nodes, configuring the preview device to 176x144 (so that
VFPP is fast) and then copying the preview stream to the encoder.

NOTE: VFPPLimitedResolutionList must be set to a proper value
for the platform in the camera_profiles.xml. If e.g. for resolution
1024*768 the FPS drops to half the normal because VFPP is too slow
to process the frame during blanking time, add 1024x768 to the XML
resolution list.

Change-Id: I14fe0241603958def7d20e01205df112c614ff23
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Signed-off-by: Juha Alanen <juha.alanen@intel.com>
Reviewed-on: http://android.intel.com:8080/98982
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>