[PORT FROM R42]libcamera2: NEW SNAPSHOT BUFFER FLOW CONTROL
BZ: 97724

In order to support more complex use cases like ULL we need
to be able to  keep, for an indefinite amount of time,
snapshot buffers out of the current buffer flow.

This is currently not possible, since we ask from PictureThread
only what we need (to allocate) for the current capture and we
use it all. Not allowing other capture requests to come in until
JPEG encoding has completed.

For ULL we need to allow the 3 snapshot buffers to be held for as long
as needed while the post capture processing is ongoing  and then
recyle the buffers.

In order to achieve this we change the control of the flow for
snapshot buffers to make it more flexible.

Allocation is still done in the PictureThread asynchronously
ControlThread is responsible for requesting enough buffers for
the current use cases (allowing for the ones that can be held).

New: snaphot buffers are not fetched when they are needed during
the take picture process (inside SetExternalSnapshotBuffers),
now they arrive via callback once PictureThread has completed the
allocation. This removes one context switch during the
take_pircture process.

Snapshot buffers are managed by 2 vectors in ControlThread:
- mAllocatedSnapshotBuffers tracks  what was  requested  from
  PictureThread. This avoid unnecessary re-allocation messages towards
  PictureThread. This list is only empty at the begining, and while
  an allocation request is in progress.
- mAvailableSnapshotBuffers is used to recycle buffer during capture
  sequence. This vector is updated with new buffers when they arrive
  from PictureThread. Also when capture is stoped the Allocated buffers
  are reset into the available buffers vector
  In a normal capture sequence ControlThread will take buffers from this
  list and set them to AtomISP before capture.
  When a capture is completed the buffer is returned to this list
  in PictureDone.
  It is responsibility of the ControlThread to ensure that there will
  be enough buffers.

Orig-Change-Id: I23f877ee21839cc0e2270d0c04d084e80eed7f1b
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Change-Id: I4d1bdcca07cd2f09ec6835e6fc622ab0aadfff41
Reviewed-on: http://android.intel.com:8080/101489
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>