libcamera2: remove redundant vector mUnqueuedPicBuf
BZ: 97724

This vector was used to store the picture buffers during
a burst and allow longer burst capture than allocated snapshot
buffers. This is no longer the case.

Besides the new vectors introduced in this patchset could
be used for the same purpose one Postview buffers are allocated
by PictureThread

Change-Id: Ida978ef998cdb996b9a74f7f783f960f4037d10b
Orig-Change-Id: I830917282ca91159be662257d3551f3660df19f9
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/101491
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>