libcamera2: set proper memory type flag for video recording buffers
BZ: 99092

We are requesting uncached memory for video recording buffers, but
the allocated buffers are actually cached.

We need to set the proper memory type flag in IntelMetaDataBuffer.
Otherwise encoded video will have corruption if the buffer contents are
modified in HAL before encoding (e.g. in case of mirroring).

Change-Id: Ia35ee5d2a132d79c7f3f399e41b90c51a994cde4
Signed-off-by: Juha Alanen <juha.alanen@intel.com>
Reviewed-on: http://android.intel.com:8080/101261
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>