libcamera2: protect against parameter changes during capturing
BZ: 99739

This patch adds capture substates which are then used to protect
against applications that erroneously try to change parameters in
the middle of capturing. Such parameter changes will result in
undeterministic results, so this patch adds en error print and
drops such parameters which are happening during the capturing.

Change-Id: I264ee787792c42678e0620ec1f79f4138a095ffd
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Reviewed-on: http://android.intel.com:8080/101716
Reviewed-by: Mandy, Arnaud <arnaud.mandy@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>