libcamera2: add more checking for flash mode in the validateParameters
BZ: 97839

Issue:
some camera application will set the "off" flash mode to the camera HAL,
even it's front camera and we dont support "off" flash mode.
This will cause the set parameter fail.
So we need to add more checking for flash mode in the validateParameters.

Solution:
1. add the checking if the sensor is back sensor and if it's support flash.
2. remove the checking in the getDefaultParameters which is in the AtomISP.cpp,
all the settings should be got from the configuration file.

Depends on:
104015 104017 104018

Change-Id: I7476424641fc9cb62fd6b893ace9d872196c15dd
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/102364
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Alanen, Juha <juha.alanen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>