libcamera2: correctly lock and unlock graphic preview buffers
BZ: 93617

We were only locking the preview buffers dequeued from the window
once, when we fetch them. And unlocking the buffers  only when we
close the PreviewThread.

We should unlock them everytime we queue them back to the window
since the unlock call triggers the cache flushes we need
to avoid the line we can see sometimes.

Change-Id: I8b033cd386501c6c80b7df3de778b23f404cd922
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/102603
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Reviewed-by: Mandy, Arnaud <arnaud.mandy@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>