libcamera2: re-init ISP during continuous capture mode
BZ: 99986

Sometimes the ISP fails and we get a error notified
it exists a timeout routine to restart the ISP.
But on continuous capture the timeout is too long
and it is required to re-init the ISP as soon as the app
requests preview.
This patch, simply checks if the ISP is initialized before
trying to start core preview.

Change-Id: Iab2b4365a55c8519e1144d1dd7d34d38f92a7ee9
Signed-off-by: Arnaud Mandy <arnaud.mandy@intel.com>
Reviewed-on: http://android.intel.com:8080/103802
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>