[PORT FROM R42] libcamera2: avoid libva creation/termination during runtime
BZ: 102232

Problem:
for unknown reasons dlclose of libVA backend library
seems to trigger the closure of  the preview video
device /dev/video2 in some random situations

In current implementation of HW JPEG encoder we re-create
the whole libVA session everytime a new snapshot resolution is
selected. This implies calls to vaInitialize and vaTerminate

Strictly speaking we only need to re-create the VA context when
the buffers change.

Workaround:
reduce the rate at which we call vaInitialize and vaTerminate
that trigger the dlopen/dlclose that seems to be causing this
problem.
We move those 2 calls from the init/deinit calls to the
constructor/destructor of the HW JPEG encoder class

Change-Id: I3bde98060854e9b8be461bbe940cf108d6308e3e
Orig-Change-Id: Ie5729636963a7bc005ccc9b9a6bfbdce49267836
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/104938
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>